+++ src/main/java/com/github/games647/fastlogin/PlayerSession.java
@@ -7,13 +7,28 @@ package com.github.games647.fastlogin;
  */
 public class PlayerSession {
 
-    private final byte[] verifyToken;
     private final String username;
+    private final String serverId;
+    private final byte[] verifyToken;
     private boolean verified;
 
-    public PlayerSession(byte[] verifyToken, String username) {
+    public PlayerSession(String username, String serverId, byte[] verifyToken) {
         this.username = username;
+        this.serverId = serverId;
         this.verifyToken = verifyToken;
+    }
+
+    /**
+     * Gets the random generated server id. This makes sure the request
+     * sent from the client is just for this server.
+     *
+     * See this for details
+     * http://www.sk89q.com/2011/09/minecraft-name-spoofing-exploit/
+     *
+     * @return random generated server id
+     */
+    public String getServerId() {
+        return serverId;
     }
 
     /**